+++ src/main/java/com/example/auction/domain/bid/repository/BidRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface BidRepository extends JpaRepository<Bid, Long>, BidCustomRepository {
@@ -17,5 +18,5 @@ public interface BidRepository extends JpaRepository<Bid, Long>, BidCustomReposi
     Page<Bid> findAllByAuctionId(Long auctionId, Pageable pageable);
 
     @Query("SELECT MIN(b.price) FROM Bid b WHERE b.auctionId = :auctionId")
-    Optional<Long> findMinPriceByAuctionId(@Param("auctionId") Long auctionId);
+    Optional<BigDecimal> findMinPriceByAuctionId(@Param("auctionId") Long auctionId);
 }
